requirements.txt
weaviate-client
weaviate
crewai
langchain-groq
langgraph
langchain-community
python-dotenv
tavily-python
streamlit